# src/docs/swagger.yaml
definitions:
...
  dto.ReqCreateUpdateProject:
    properties:
      about:
        type: string
      color:
        type: string
      id:
        type: integer
      is_private:
        type: boolean
      name:
        type: string
    required:
    - name
    type: object
...
  dto.RespProject:
    properties:
      about:
        type: string
      color:
        type: string
      id:
        type: integer
      is_private:
        type: boolean
      name:
        type: string
      user_id:
        type: integer
    type: object
...
  echo.HTTPError:
    properties:
      message: {}
    type: object
...
  pkg.Response:
    properties:
      body: {}
    type: object
info:
  contact: {}
paths:
...
  /project/{id}:
    delete:
      consumes:
      - application/json
      description: Delete an project
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "404":
          description: can't find project with such id
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Delete an project
      tags:
      - project
    get:
      consumes:
      - application/json
      description: Get project by id
      parameters:
      - description: Project ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: success get project
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespProject'
              type: object
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Show a post
      tags:
      - project
  /project/create:
    post:
      consumes:
      - application/json
      description: Create project
      parameters:
      - description: project info
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateUpdateProject'
      produces:
      - application/json
      responses:
        "200":
          description: success update project
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespProject'
              type: object
        "400":
          description: bad req
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf or permission denied
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "422":
          description: unprocessable entity
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Create project
      tags:
      - project
  /project/edit:
    post:
      consumes:
      - application/json
      description: Update an project
      parameters:
      - description: project info
        in: body
        name: project
        required: true
        schema:
          $ref: '#/definitions/dto.ReqCreateUpdateProject'
      produces:
      - application/json
      responses:
        "200":
          description: success update project
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  $ref: '#/definitions/dto.RespProject'
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "403":
          description: invalid csrf or permission denied
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: invalid http method
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "422":
          description: unprocessable entity
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Update an project
      tags:
      - project
  /project/usr:
    get:
      description: Get user projects or get user projects for a day
      produces:
      - application/json
      responses:
        "200":
          description: success get projects
          schema:
            allOf:
            - $ref: '#/definitions/pkg.Response'
            - properties:
                body:
                  items:
                    $ref: '#/definitions/dto.RespProject'
                  type: array
              type: object
        "400":
          description: bad request
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "401":
          description: no cookie
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "405":
          description: Method Not Allowed
          schema:
            $ref: '#/definitions/echo.HTTPError'
        "500":
          description: internal server error
          schema:
            $ref: '#/definitions/echo.HTTPError'
      summary: Get user projects
      tags:
      - project
...
swagger: "2.0"